+++ apps/vscode-extension/src/services/index.ts
@@ -2,6 +2,7 @@ export * from './analytics';
 export * from './ChatActionsService';
 export * from './DecoratorService';
 export * from './DemoApi';
+export * from './DesktopActionsService';
 export * from './DemoCreator';
 export * from './DemoFileProvider';
 export * from './DemoListeners';